+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
--
+### Added
+
+- Allow download multiple cover sizes at once. Each cover size must be provided with the `--cover-size` option
 
 ## [0.2.3] - 2022-09-06